+++ mean.h
@@ -1,7 +1,7 @@
 #ifndef MEAN_H
 #define MEAN_H
 
-float mean(int num[], int n);
+float mean(float num[], int n);
 
 
 #endif
